requirements.txt
torch
transformers
pillow
ollama

